requirements.txt
Django==1.8
django-mptt==0.7.1
django-rest-auth==0.3.4
djangorestframework==3.1.1
ecdsa==0.13
Fabric==1.10.1
gunicorn==19.3.0
paramiko==1.15.2
pycrypto==2.6.1
six==1.9.0
